requirements.txt
flask
flask-caching
azure-search-documents
openai
python-dotenv
pyodbc